ViewSpecificClassBindableBase moved to its own file to simplify its comparing with ViewSpecificClassBase
